backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.1
supabase>=2.9.0
sentence-transformers>=3.2.0
faiss-cpu>=1.8.0
PyMuPDF>=1.24.0
python-docx>=1.1.2
google-genai>=0.3.0
numpy>=1.26.0
pydantic>=2.9.0
python-multipart>=0.0.12
firebase-admin>=6.5.0
pytest>=8.3.0
slowapi>=0.1.9
rank-bm25>=0.2.2
langchain-text-splitters>=0.3.0
cachetools>=5.5.0
